+++ hooks/usebreadcrumbs.jsx
@@ -1,39 +1,22 @@
-﻿import { useRouter } from 'next/router';
-import sections from 'components/combindesections'; 
 import site from 'config/site';
-
+import { useRouter } from 'next/router';
 
 const useBreadcrumbs = () => {
-  const router = useRouter(); 
-  const lang  = router == 'en' ? 'en' : 'no'
-
+  const router = useRouter();
   const pathSegments = router.pathname.split('/').filter(segment => segment);
 
-  return pathSegments.map((segment, index ) => {
-    var href = '/' + pathSegments.slice(0, index + 1).join('/').replace(/\.(no|en)$/, '');
-    var seg = segment.replace(/\.(no|en)$/, '')  ;
-    var label = seg
-    var folder = ''
+  return pathSegments.map((segment, index) => {
+    let href = '/' + pathSegments.slice(0, index + 1).join('/').replace(/\.(no|en)$/, '');
+    let label = segment.replace(/\.(no|en)$/, '');
 
-    if (index == 0 ) {
-      label = site?.bolkene?.items[seg]?.title[lang]  ?? '…'
-      href =  site?.bolkene?.items[seg]?.href ?? '';
-      
+    if (index == 0) {
+      href = site.bolkene.items[label].href;
+      label = site.bolkene.items[label].title.no
     }
 
-    
-    if (index == 1 ) {
-      // get title by looking it up from 'combinedsections'
-      const bolk = pathSegments.slice(0, 1)[0];
-      label = sections[bolk]?.items[label]?.title[lang] ??   '…'
-      folder = sections[bolk]?.items;
-      // return { href, label, index, folder: sections[bolk].items, key: seg+label+index  };
-     }
-     return { href, label, index, folder, key: seg+label+index  };
+    //TO DO: Check language if we use the title
+    return { href, label, index };
   });
 };
 
 export default useBreadcrumbs;
-
-
-
